Move 'blurb test' subcommand into test suite
