+++ tox.ini
@@ -17,9 +17,7 @@ commands =
       --cov-report term \
       --cov-report xml \
       {posargs}
-    blurb test
     blurb help
     blurb --version
-    {envpython} -I -m blurb test
     {envpython} -I -m blurb help
     {envpython} -I -m blurb version
